# src/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Secure Authentication API",
    "version": "1.0.0",
    "description": "A secure REST API with Supabase Auth, JWT verification, and protected routes"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 6
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {"type": "string"},
                        "email": {"type": "string"},
                        "created_at": {"type": "string"}
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing fields or invalid data"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login with email and password",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {"type": "string"},
                    "refresh_token": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "400": {
            "description": "Bad request - missing fields"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout current user",
        "tags": ["Authentication"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "204": {
            "description": "Logout successful"
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get user profile",
        "tags": ["Protected"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "string"},
                    "email": {"type": "string"},
                    "created_at": {"type": "string"}
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get protected dashboard",
        "tags": ["Protected"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dashboard data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"},
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {"type": "string"},
                        "email": {"type": "string"},
                        "created_at": {"type": "string"}
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid or missing token"
          }
        }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Get public information",
        "tags": ["Public"],
        "responses": {
          "200": {
            "description": "Public information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}